# swagger.yaml
openapi: 3.0.3
info:
  title: Branded UK E-Commerce API
  description: |
    # Overview
    This API provides endpoints for the Branded UK e-commerce platform.
    
    ## Base URL
    **Production**: `https://api.brandeduk.com`
    
    ## Authentication
    No authentication required. Public API.
    
    ## Endpoint Categories
    
    ### Products (Main)
    - `GET /api/products` - Main product listing with combined filters
    - `GET /api/products/filters` - Product type counts for filtering
    - `GET /api/products/types` - All product types with counts
    - `GET /api/products/{code}` - Single product details
    - `GET /api/products/{code}/pricing` - Get pricing information for a style code
    
    ### Contact & Quotes
    - `POST /api/contact` - Submit contact form inquiry
    - `POST /api/quotes` - Submit quote request with optional logo file uploads
    
    ### Filter Endpoints (Testing & Database Exploration)
    Each filter type has TWO endpoints:
    1. **List all values**: Returns all options from that database table
    2. **Get products by slug**: Returns products matching a specific filter value
    
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://api.brandeduk.com
    description: Production Server
  - url: http://localhost:3004
    description: Local Development
...
paths:
...
  /api/products:
    get:
      tags:
        - Products
      summary: Get product list (main combined filter endpoint)
      description: |
        Returns a paginated list of products with optional filtering and search.
        This is the main endpoint that supports all filters combined.
      operationId: getProducts
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Items per page (max 200)
          schema:
            type: integer
            default: 24
        - name: q
          in: query
          description: Search query
          schema:
            type: string
        - name: priceMin
          in: query
          schema:
            type: number
        - name: priceMax
          in: query
          schema:
            type: number
        - name: productType
          in: query
          description: Product type slug(s)
          schema:
            type: string
        - name: gender
          in: query
          schema:
            type: string
        - name: ageGroup
          in: query
          schema:
            type: string
        - name: sleeve
          in: query
          schema:
            type: string
        - name: neckline
          in: query
          schema:
            type: string
        - name: fabric
          in: query
          schema:
            type: string
        - name: size
          in: query
          schema:
            type: string
        - name: colour
          in: query
          schema:
            type: string
        - name: primaryColour
          in: query
          schema:
            type: string
        - name: style
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: weight
          in: query
          schema:
            type: string
        - name: fit
          in: query
          schema:
            type: string
        - name: sector
          in: query
          schema:
            type: string
        - name: sport
          in: query
          schema:
            type: string
        - name: effect
          in: query
          schema:
            type: string
        - name: accreditations
          in: query
          schema:
            type: string
        - name: brand
          in: query
          description: Brand slug(s) or name(s) - accepts multiple values
          schema:
            type: string
          example: "awdis"
        - name: brands
          in: query
          description: Alias for brand parameter (accepts multiple values)
          schema:
            type: string
        - name: colourShade
          in: query
          schema:
            type: string
        - name: sort
          in: query
          description: |
            Sort option. Available values:
            - `newest` - Sort by creation date (newest first)
            - `best` - Best sellers (uses newest as proxy)
            - `brand-az` - Sort by brand name A-Z
            - `brand-za` - Sort by brand name Z-A
            - `code-az` - Sort by product code A-Z
            - `code-za` - Sort by product code Z-A
            - `price-lh` - Sort by price low to high
            - `price-hl` - Sort by price high to low
            - `price` - Sort by price (use with order parameter)
            - `name` - Sort by product name (use with order parameter)
          schema:
            type: string
            enum: [newest, best, brand-az, brand-za, code-az, code-za, price-lh, price-hl, price, price-low, price-high, name]
            default: newest
            default: newest
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  priceRange:
                    type: object
                    properties:
                      min:
                        type: number
                      max:
                        type: number
  /api/products/filters:
    get:
      tags:
        - Products
      summary: Get filter aggregations (counts)
      description: Returns counts for all filter types based on current filters
      operationId: getProductFilters
      parameters:
        - name: q
          in: query
          description: Search query
          schema:
            type: string
        - name: productType
          in: query
          schema:
            type: string
        - name: gender
          in: query
          schema:
            type: string
        - name: ageGroup
          in: query
          schema:
            type: string
        - name: sleeve
          in: query
          schema:
            type: string
        - name: neckline
          in: query
          schema:
            type: string
        - name: fabric
          in: query
          schema:
            type: string
        - name: size
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  filters:
                    type: object
                    properties:
                      gender:
                        type: object
                        description: Map of gender slug to count
                      ageGroup:
                        type: object
                        description: Map of age group slug to count
                      sleeve:
                        type: object
                        description: Map of sleeve slug to count
                      neckline:
                        type: object
                        description: Map of neckline slug to count
                      fabric:
                        type: object
                        description: Map of fabric slug to count
                      size:
                        type: object
                        description: Map of size slug to count
                      feature:
                        type: object
                        description: Map of feature slug to count
                      tag:
                        type: object
                        description: Map of tag slug to count
...
  /api/display-order:
    get:
      tags:
        - Display Order
      summary: List display order entries
      description: |
        Returns paginated list of `product_display_order` entries.
        You can optionally filter by `brand_id`, `product_type_id` and/or `style_code`.
      parameters:
        - name: brand_id
          in: query
          schema:
            type: integer
          description: Filter by brand ID
        - name: product_type_id
          in: query
          schema:
            type: integer
          description: Filter by product type ID
        - name: style_code
          in: query
          schema:
            type: string
          description: Filter by style code (partial match)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DisplayOrderEntry'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Display Order
      summary: Create or update a single display order entry
      description: |
        Upserts a `product_display_order` row for a specific style code and optional brand/product type context.
        Context uniqueness is enforced on `(style_code, brand_id, product_type_id)`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - style_code
                - display_order
              properties:
                style_code:
                  type: string
                  description: Product style code (e.g. "JC001")
                  example: "JC001"
                brand_id:
                  type: integer
                  nullable: true
                  description: Optional brand context ID
                  example: 28
                product_type_id:
                  type: integer
                  nullable: true
                  description: Optional product type context ID
                  example: 21
                display_order:
                  type: integer
                  description: Lower numbers appear first (1 = first)
                  example: 1
      responses:
        '201':
          description: Display order saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Display order saved successfully
                  data:
                    $ref: '#/components/schemas/DisplayOrderEntry'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Style code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/display-order/products:
    get:
      tags:
        - Display Order
      summary: Get products with display order for admin
      description: |
        Returns all products for a brand and/or product type including their current
        `display_order` value, for use in the admin sorting interface.
      parameters:
        - name: brand_id
          in: query
          schema:
            type: integer
          description: Filter products by brand ID
        - name: product_type_id
          in: query
          schema:
            type: integer
          description: Filter products by product type ID
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DisplayOrderProduct'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                  filters:
                    type: object
                    properties:
                      brand_id:
                        type: integer
                        nullable: true
                      product_type_id:
                        type: integer
                        nullable: true
        '400':
          description: Missing required brand_id or product_type_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/categories:
    get:
      tags:
        - Categories
      summary: Get all categories
      description: |
        Returns all categories with hierarchical structure (parent categories with subcategories).
        Supports both hierarchical and flat formats.
      operationId: getAllCategories
      parameters:
        - name: format
          in: query
          description: Response format
          schema:
            type: string
            enum: [hierarchical, flat]
            default: hierarchical
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/contact:
    post:
      tags:
        - Contact
      summary: Submit contact form
      description: |
        Submit a contact form inquiry. Supports rate limiting (max 3 submissions per email per hour).
        Sends email notification to the configured recipient address.
      operationId: submitContactForm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactFormRequest'
            example:
              name: "John Doe"
              email: "john@example.com"
              interest: "embroidery"
              phone: "+44 20 1234 5678"
              address: "123 Main Street, London"
              postCode: "SW1A 1AA"
              message: "I need custom workwear for my team of 50 employees."
      responses:
        '200':
          description: Contact form submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactFormResponse'
              example:
                success: true
                message: "Thank you for your inquiry. We'll get back to you soon."
                id: "contact_1769198076999_x80gdu004"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactFormError'
              example:
                success: false
                error: "Validation failed"
                errors:
                  email: "Invalid email format"
                  name: "Name is required"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Too many submissions"
                  message:
                    type: string
                    example: "Maximum 3 contact form submissions per hour. Please try again later."
                  retryAfter:
                    type: integer
                    example: 1800
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================================================
  # QUOTE REQUEST ENDPOINT
  # ============================================================================
  /api/quotes:
    post:
      tags:
        - Quotes
      summary: Submit quote request with optional logo uploads
      description: |
        Submit a quote request for custom products. Supports two formats:
        1. **multipart/form-data** - When logo files are included (recommended)
        2. **application/json** - Backward compatible format without files
        
        Logo files are named using the pattern: `logo_{position-slug}` (e.g., `logo_left-breast`, `logo_large-front-center`).
        Maximum 10 logo files, 5MB each. Only image files (jpg, png, gif, webp) are accepted.
      operationId: submitQuoteRequest
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - quoteData
              properties:
                quoteData:
                  type: string
                  description: JSON string containing all quote data (customer, basket, customizations, summary, timestamp)
                  example: '{"customer":{"fullName":"John Doe","email":"john@example.com","phone":"+44 20 1234 5678","company":"Acme Corp","address":"123 Main St"},"basket":[{"name":"Product Name","code":"PROD-123","color":"Navy","quantity":50,"sizes":{"S":10,"M":20,"L":20},"unitPrice":10.00,"itemTotal":500.00}],"customizations":[{"position":"Left Breast","method":"Embroidery","type":"logo","hasLogo":true,"unitPrice":5.00,"lineTotal":250.00,"quantity":50}],"summary":{"totalQuantity":50,"totalItems":1,"garmentCost":500.00,"customizationCost":250.00,"digitizingFee":25.00,"subtotal":775.00,"vatRate":0.20,"vatAmount":155.00,"totalExVat":775.00,"totalIncVat":930.00,"vatMode":"inc","displayTotal":930.00},"timestamp":"2026-01-15T10:30:00.000Z"}'
                logo_left-breast:
                  type: string
                  format: binary
                  description: Logo image file for left breast position (max 5MB, jpg/png/gif/webp)
                logo_right-breast:
                  type: string
                  format: binary
                  description: Logo image file for right breast position (max 5MB, jpg/png/gif/webp)
                logo_large-front-center:
                  type: string
                  format: binary
                  description: Logo image file for large front center position (max 5MB, jpg/png/gif/webp)
                logo_small-centre-front:
                  type: string
                  format: binary
                  description: Logo image file for small centre front position (max 5MB, jpg/png/gif/webp)
                logo_back-center:
                  type: string
                  format: binary
                  description: Logo image file for back center position (max 5MB, jpg/png/gif/webp)
                logo_left-sleeve:
                  type: string
                  format: binary
                  description: Logo image file for left sleeve position (max 5MB, jpg/png/gif/webp)
                logo_right-sleeve:
                  type: string
                  format: binary
                  description: Logo image file for right sleeve position (max 5MB, jpg/png/gif/webp)
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
            example:
              customer:
                fullName: "John Doe"
                company: "Acme Corp"
                phone: "+44 20 1234 5678"
                email: "john@example.com"
                address: "123 Main St, London"
              summary:
                totalQuantity: 50
                totalItems: 1
                garmentCost: 500.00
                customizationCost: 250.00
                digitizingFee: 25.00
                subtotal: 775.00
                vatRate: 0.20
                vatAmount: 155.00
                totalExVat: 775.00
                totalIncVat: 930.00
                vatMode: "inc"
                displayTotal: 930.00
              basket:
                - name: "Product Name"
                  code: "PROD-123"
                  color: "Navy"
                  quantity: 50
                  sizes:
                    S: 10
                    M: 20
                    L: 20
                  sizesSummary: "S: 10, M: 20, L: 20"
                  unitPrice: 10.00
                  itemTotal: 500.00
              customizations:
                - position: "Left Breast"
                  method: "Embroidery"
                  type: "logo"
                  hasLogo: true
                  unitPrice: 5.00
                  lineTotal: 250.00
                  quantity: 50
              timestamp: "2026-01-15T10:30:00.000Z"
      responses:
        '200':
          description: Quote submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
              example:
                success: true
                message: "Quote submitted successfully"
                quoteId: "quote_1769198120482_i98a3pwgg"
        '400':
          description: Validation error or file upload error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    examples:
                      - "Customer email is required"
                      - "Customer full name is required"
                      - "Customer phone is required"
                      - "Basket is required"
                      - "File too large. Maximum size is 5MB per file."
                      - "Too many files. Maximum 10 logo files allowed."
                      - "Only image files are allowed."
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        parentId:
          type: integer
          nullable: true
        displayOrder:
          type: integer
        description:
          type: string
          nullable: true
        categoryType:
          type: string
          nullable: true
        subcategories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
    ContactFormRequest:
      type: object
      required:
        - name
        - email
        - interest
        - message
      properties:
        name:
          type: string
          description: Contact's full name (2-100 characters)
          example: "John Doe"
        email:
          type: string
          format: email
          description: Valid email address
          example: "john@example.com"
        interest:
          type: string
          enum: [embroidery, printing, workwear, uniforms, promotional, other]
          description: What they're interested in
          example: "embroidery"
        phone:
          type: string
          description: Phone number (optional, 7-20 characters)
          example: "+44 20 1234 5678"
        address:
          type: string
          description: Street address (optional, desktop form only)
          example: "123 Main Street, London"
        postCode:
          type: string
          description: UK post code (optional, desktop form only)
          example: "SW1A 1AA"
        message:
          type: string
          description: User's message/query (10-5000 characters)
          example: "I need custom workwear for my team of 50 employees."
    ContactFormResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Thank you for your inquiry. We'll get back to you soon."
        id:
          type: string
          description: Unique contact submission ID
          example: "contact_1769198076999_x80gdu004"
    ContactFormError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation failed"
        errors:
          type: object
          additionalProperties:
            type: string
          example:
            email: "Invalid email format"
            name: "Name is required"
            interest: "Interest is required"
            message: "Message must be at least 10 characters"
...
    QuoteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Quote submitted successfully"
        quoteId:
          type: string
          description: Unique quote submission ID
          example: "quote_1769198120482_i98a3pwgg"
    DisplayOrderEntry:
      type: object
      description: Single product_display_order row with joined metadata
      properties:
        id:
          type: integer
        style_code:
          type: string
          example: "YP047"
        brand_id:
          type: integer
          nullable: true
        brand_name:
          type: string
          nullable: true
        product_type_id:
          type: integer
          nullable: true
        product_type_name:
          type: string
          nullable: true
        display_order:
          type: integer
          description: Lower numbers appear first (1 = first)
          example: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        product_name:
          type: string
          nullable: true
          description: Name of the style from styles table
...
    DisplayOrderProduct:
      type: object
      description: Product row as returned from /api/display-order/products
      properties:
        style_code:
          type: string
          example: "YP047"
        product_name:
          type: string
          example: "Flexfit Classic"
        brand_id:
          type: integer
        brand_name:
          type: string
        product_type_id:
          type: integer
        product_type_name:
          type: string
        display_order_id:
          type: integer
          nullable: true
          description: Underlying product_display_order.id if it exists
        display_order:
          type: integer
          description: Effective display order (defaults to 999999 when not set)
        display_order_created_at:
          type: string
          format: date-time
          nullable: true
        image:
          type: string
          nullable: true
          description: Primary product image URL
...